move profile avatar to new UserAvatar component
